move graceful thread shutdown to a helper
