+++ airbyte-commons/src/main/java/io/airbyte/commons/concurrency/GracefulShutdownHandler.java
@@ -22,19 +22,22 @@
  * SOFTWARE.
  */
 
-package io.airbyte.scheduler;
+package io.airbyte.commons.concurrency;
 
+import java.time.Duration;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SchedulerShutdownHandler extends Thread {
+public class GracefulShutdownHandler extends Thread {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(SchedulerShutdownHandler.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(GracefulShutdownHandler.class);
+  private final Duration terminateWaitDuration;
   private final ExecutorService[] threadPools;
 
-  public SchedulerShutdownHandler(final ExecutorService... threadPools) {
+  public GracefulShutdownHandler(Duration terminateWaitDuration, final ExecutorService... threadPools) {
+    this.terminateWaitDuration = terminateWaitDuration;
     this.threadPools = threadPools;
   }
 
@@ -44,11 +47,11 @@ public class SchedulerShutdownHandler extends Thread {
       threadPool.shutdown();
 
       try {
-        if (!threadPool.awaitTermination(30, TimeUnit.SECONDS)) {
-          LOGGER.error("Unable to kill worker threads by shutdown timeout.");
+        if (!threadPool.awaitTermination(terminateWaitDuration.getSeconds(), TimeUnit.SECONDS)) {
+          LOGGER.error("Unable to kill threads by shutdown timeout.");
         }
       } catch (InterruptedException e) {
-        LOGGER.error("Wait for graceful worker thread shutdown interrupted.", e);
+        LOGGER.error("Wait for graceful thread shutdown interrupted.", e);
       }
     }
   }
